+++ algorithm/include/search.hpp
@@ -22,9 +22,9 @@
 namespace algorithm::search
 {
 //! @brief Maximum alignment length per element of printing.
-constexpr uint32_t maxAlignOfPrint = 16;
+constexpr uint8_t maxAlignOfPrint = 16;
 //! @brief Maximum columns per row of printing.
-constexpr uint32_t maxColumnOfPrint = 10;
+constexpr uint8_t maxColumnOfPrint = 10;
 
 //! @brief Solution of search.
 //! @tparam T - type of the solution of search
