+++ utility/source/hash.cpp
@@ -11,10 +11,10 @@ namespace utility::hash
 //! @brief The Brian-Kernighan Dennis-Ritchie hash function.
 //! @param str - input data
 //! @return hash value
-uint32_t bkdrHash(const char* str)
+uint64_t bkdrHash(const char* str)
 {
-    const uint32_t seed = bkdrHashSeed; // 31 131 1313 13131 131313 etc...
-    uint32_t hash = 0;
+    const uint64_t seed = bkdrHashSeed; // 31 131 1313 13131 131313 etc...
+    uint64_t hash = 0;
     while (*str)
     {
         hash = hash * seed + (*str++);
